Move modal components to its own group directory
